Move map to MapProjectGeometry as third step of New Project
